In a half open drawer, show types without an accordion
May help new users figure out how to close the drawer
