+++ src/components/entry/EntryOverview.js
@@ -14,6 +14,7 @@ import styled from 'styled-components/macro'
 import { MIN_LOCATION_ZOOM } from '../../constants/map'
 import { setPaneDrawerToLowPosition } from '../../redux/locationSlice'
 import { useAppHistory } from '../../utils/useAppHistory'
+import { useIsDesktop } from '../../utils/useBreakpoint'
 import { theme } from '../ui/GlobalStyle'
 import IconBesideText from '../ui/IconBesideText'
 import { ReportButton } from './overview/ReportButton'
@@ -77,6 +78,7 @@ const EntryOverview = () => {
   const { locationsWithoutPanorama } = useSelector((state) => state.misc)
   const user = useSelector((state) => state.auth.user)
   const dispatch = useDispatch()
+  const isDesktop = useIsDesktop()
 
   const { t, i18n } = useTranslation()
 
@@ -113,7 +115,10 @@ const EntryOverview = () => {
   return (
     <div>
       <>
-        <TypesHeader types={types} />
+        <TypesHeader
+          types={types}
+          openable={pane.drawerFullyOpen || isDesktop}
+        />
         <Tags locationData={locationData} />
         <Description>
           <p dir="auto">{locationData.description}</p>
